fix(RCSwitch): read micros() unsigned in the receive interrupt
handleInterrupt() took micros() into a long and subtracted lastTime from it
there. micros() passes what a signed 32-bit value holds after about 36 minutes
of uptime, so the conversion is implementation-defined from then on and the
subtraction that follows is a signed overflow.

Both are unsigned now, which is where wrapping is defined and what diff()
already does with the same durations. RCSwitch is lib_ignored in native_test
and stands in as a shim, so the interrupt handler is not reachable from a host
test.

+++ lib/RCSwitch/src/RCSwitch.cpp
@@ -457,7 +457,9 @@ void RCSwitch::handleInterrupt() {
   static uint32_t changeCount = 0;
   static uint32_t lastTime = 0;
 
-  const long time = micros();
+  // Unsigned, like diff(): micros() passes what a signed 32-bit value holds after about 36
+  // minutes, and the subtraction below only stays defined while both sides are unsigned.
+  const uint32_t time = micros();
   const uint32_t duration = time - lastTime;
 
   RCSwitch::buftimings[3] = RCSwitch::buftimings[2];
